+++ src/state/scomplexlist.cc
@@ -266,9 +266,9 @@ double SComplexList::getJoinFlux(void) {
 
 }
 
-uint16_t SComplexList::getMoveCount(void) {
+int SComplexList::getMoveCount(void) {
 
-	uint16_t output = 0;
+	int output = 0;
 
 	for (SComplexListEntry* temp = first; temp != NULL; temp = temp->next) {
 
@@ -804,8 +804,8 @@ bool SComplexList::checkLooseStructure(const char *our_struc, const char *stop_s
 	intvec our_pairs, stop_pairs;
 	int remaining_distance = count;
 
-	len = strlen(our_struc);
-	if (len != strlen(stop_struc))
+	len = (int)strlen(our_struc);
+	if (len != (int)strlen(stop_struc))
 		return false;  // something weird happened, as it should have the
 // same ID list...
 
@@ -862,8 +862,8 @@ bool SComplexList::checkCountStructure(const char *our_struc, const char *stop_s
 	intvec our_pairs, stop_pairs;
 	int remaining_distance = count;
 
-	len = strlen(our_struc);
-	if (len != strlen(stop_struc))
+	len = (int)strlen(our_struc);
+	if (len != (int)strlen(stop_struc))
 		return false;  // something weird happened, as it should have the
 // same ID list...
